+++ test-parts/01-common_data/items/itm-common_objects.js
@@ -4,6 +4,7 @@ const commonPaths = require("../../../app/paths/files/app-paths");
 const commonObjectsFile = require(commonPaths.commonObjects);
 const testCommon = require(commonPaths.testCommon);
 const testObject = require(commonPaths.testObject);
+const testArray = require(commonPaths.testArray);
 
 
 function testCommonObjects()
@@ -39,17 +40,6 @@ function handleDeviceObject(devicePropName)
 	
 	testCommon.testObject(deviceObj);
 	
-	testObjectProperty(deviceObj, 'id', 'string');
-	testObjectProperty(deviceObj, 'deviceType', 'string');
-	testObjectProperty(deviceObj, 'maker', 'string');
-	testObjectProperty(deviceObj, 'model', 'string');
-	testObjectProperty(deviceObj, 'name', 'string');
-	testObjectProperty(deviceObj, 'desc', 'string');
-	testObjectProperty(deviceObj, 'ipAddress', 'string');
-	testObjectProperty(deviceObj, 'username', 'string');
-	testObjectProperty(deviceObj, 'password', 'string');
-	testObjectProperty(deviceObj, 'isEnabled', 'boolean');
-	
 	testCommon.testString(deviceObj.id);
 	testCommon.testString(deviceObj.deviceType);
 	testCommon.testString(deviceObj.maker);
@@ -59,6 +49,7 @@ function handleDeviceObject(devicePropName)
 	testCommon.testString(deviceObj.ipAddress);
 	testCommon.testString(deviceObj.username);
 	testCommon.testString(deviceObj.password);
+	testCommon.testBoolean(deviceObj.isEnabled);
 }
 
 function handleRegisterNodeObject()
@@ -71,27 +62,14 @@ function handleRegisterNodeObject()
 	
 	testCommon.testObject(registerResult);
 	
-	testObjectProperty(registerResult, 'id', 'string');
-	testObjectProperty(registerResult, 'type', 'string');
-	testObjectProperty(registerResult, 'z', 'string');
-	testObjectProperty(registerResult, 'name', 'string');
-	testObjectProperty(registerResult, 'deviceId', 'string');
-	testObjectProperty(registerResult, 'ioSetId', 'string');
-	testObjectProperty(registerResult, 'x', 'number');
-	testObjectProperty(registerResult, 'y', 'number');
-	testObjectProperty(registerResult, 'wires', 'array');
-	
 	testCommon.testString(registerResult.id);
 	testCommon.testString(registerResult.type);
 	testCommon.testString(registerResult.z);
 	testCommon.testString(registerResult.deviceId);
 	testCommon.testString(registerResult.ioSetId);
-}
-
-function testObjectProperty(tObject, propName, desiredType)
-{	
-	testObject.testPropExists(tObject, propName);
-	testObject.testPropType(tObject, propName, desiredType);
+	// X
+	// Y
+	testArray.testNeutral(wires.array);
 }
 
 module.exports = testCommonObjects;